+++ src/index.js
@@ -1,8 +1,9 @@
 import './style.css';
 import reloadIcon from './images/refresh.png';
 import enterIcon from './images/enter.png';
-import { refreshList } from './render.js';
-import { addItem } from './CRUD.js';
+import { enterTask, addItem, refreshList } from './CRUD.js';
+
+window.onload = refreshList;
 
 // Load images
 const refresh = document.querySelector('.refresh');
@@ -10,13 +11,10 @@ refresh.src = reloadIcon;
 const enter = document.querySelector('.enter');
 enter.src = enterIcon;
 
-const enterTask = document.querySelector('.task-input');
-document.querySelector('.enter').addEventListener('click', addItem);
+enter.addEventListener('click', addItem);
 enterTask.addEventListener('keyup', (event) => {
   event.preventDefault();
   if (event.keyCode === 13) {
     document.querySelector('.enter').click();
   }
 });
-
-refreshList();
